Module Re_posix

deprecated Use Re.Posix

include module type of struct include Re.Posix end

exception Parse_error

exception Not_supported

Errors that can be raised during the parsing of the regular expression

type opt = [

| `ICase

| `NoSub

| `Newline

]

val re : ?opts:opt list -> string -> (Re__.Ast.cset, [ `Cased | `Uncased ]) Re__.Ast.gen

Parsing of a Posix extended regular expression

val compile : (Re__.Ast.cset, [ `Cased | `Uncased ]) Re__.Ast.gen -> Re__.Compile.re

compile r is defined as Core.compile (Core.longest r)

val compile_pat : ?opts:opt list -> string -> Re__.Compile.re

compile_pat ?opts regex compiles the Posix extended regular expression regexp